relocating ARC_TILE_WORKING_AREA
