+++ src/components/shared/Board/CircularBoard/CircularBoard.tsx
@@ -9,7 +9,8 @@ import { SquareShape } from "game";
 import { objectMatches } from "utilities";
 import { CircularBackboard } from "./CircularBackboard";
 import { polarToCartesian, euclideanDistance } from "utilities";
-import { describeArc, ARC_TILE_WORKING_AREA } from "ui/Tiles/Arc";
+import { describeArc } from "ui/Tiles/Arc";
+import { ARC_TILE_WORKING_AREA } from "ui/Tiles";
 import { useCircularRotation } from "../useCircularRotation";
 import type { Point, Degrees } from "game/types";
 import { PlayerName } from "game/types";
